+++ pkg/querier/tripperware/queryrange/query_range_test.go
@@ -4,11 +4,12 @@ import (
 	"bytes"
 	"compress/gzip"
 	"context"
-	"fmt"
 	"io"
 	"net/http"
 	"strconv"
 	"testing"
+
+	"github.com/gogo/protobuf/proto"
 
 	"github.com/prometheus/common/model"
 
@@ -93,42 +94,193 @@ func TestResponse(t *testing.T) {
 	r := *parsedResponse
 	rWithWarnings := *parsedResponseWithWarnings
 	rWithInfos := *parsedResponseWithInfos
-	r.Headers = respHeaders
-	rWithWarnings.Headers = respHeaders
-	rWithInfos.Headers = respHeaders
-	for i, tc := range []struct {
-		body                  string
-		expected              *tripperware.PrometheusResponse
+	testCases := []struct {
+		promBody              *tripperware.PrometheusResponse
+		jsonBody              string
 		expectedDecodeErr     error
 		cancelCtxBeforeDecode bool
+		isProtobuf            bool
 	}{
 		{
-			body:     responseBody,
-			expected: &r,
+			promBody: &tripperware.PrometheusResponse{
+				Status: "success",
+				Data: tripperware.PrometheusData{
+					ResultType: model.ValMatrix.String(),
+					Result: tripperware.PrometheusQueryResult{
+						Result: &tripperware.PrometheusQueryResult_Matrix{
+							Matrix: &tripperware.Matrix{
+								SampleStreams: []tripperware.SampleStream{
+									{
+										Labels: []cortexpb.LabelAdapter{
+											{Name: "foo", Value: "bar"},
+										},
+										Samples: []cortexpb.Sample{
+											{Value: 137, TimestampMs: 1536673680000},
+											{Value: 137, TimestampMs: 1536673780000},
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+			jsonBody:   responseBody,
+			isProtobuf: true,
 		},
 		{
-			body:     responseBodyWithWarnings,
-			expected: &rWithWarnings,
+			promBody:   &r,
+			jsonBody:   responseBody,
+			isProtobuf: false,
 		},
 		{
-			body:     responseBodyWithInfos,
-			expected: &rWithInfos,
+			promBody: &tripperware.PrometheusResponse{
+				Status:   "success",
+				Warnings: []string{"test-warn"},
+				Data: tripperware.PrometheusData{
+					ResultType: model.ValMatrix.String(),
+					Result: tripperware.PrometheusQueryResult{
+						Result: &tripperware.PrometheusQueryResult_Matrix{
+							Matrix: &tripperware.Matrix{
+								SampleStreams: []tripperware.SampleStream{
+									{
+										Labels: []cortexpb.LabelAdapter{
+											{Name: "foo", Value: "bar"},
+										},
+										Samples: []cortexpb.Sample{
+											{Value: 137, TimestampMs: 1536673680000},
+											{Value: 137, TimestampMs: 1536673780000},
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+			jsonBody:   responseBodyWithWarnings,
+			isProtobuf: true,
 		},
 		{
-			body:                  responseBody,
+			promBody:   &rWithWarnings,
+			jsonBody:   responseBodyWithWarnings,
+			isProtobuf: false,
+		},
+		{
+			promBody: &tripperware.PrometheusResponse{
+				Status: "success",
+				Infos:  []string{"PromQL info: metric might not be a counter, name does not end in _total/_sum/_count/_bucket: \"go_gc_gogc_percent\" (1:6)"},
+				Data: tripperware.PrometheusData{
+					ResultType: model.ValMatrix.String(),
+					Result: tripperware.PrometheusQueryResult{
+						Result: &tripperware.PrometheusQueryResult_Matrix{
+							Matrix: &tripperware.Matrix{
+								SampleStreams: []tripperware.SampleStream{
+									{
+										Labels: []cortexpb.LabelAdapter{
+											{Name: "foo", Value: "bar"},
+										},
+										Samples: []cortexpb.Sample{
+											{Value: 137, TimestampMs: 1536673680000},
+											{Value: 137, TimestampMs: 1536673780000},
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+			jsonBody:   responseBodyWithInfos,
+			isProtobuf: true,
+		},
+		{
+			promBody:   &rWithInfos,
+			jsonBody:   responseBodyWithInfos,
+			isProtobuf: false,
+		},
+		{
+			promBody: &tripperware.PrometheusResponse{
+				Status: "success",
+				Data: tripperware.PrometheusData{
+					ResultType: model.ValMatrix.String(),
+					Result: tripperware.PrometheusQueryResult{
+						Result: &tripperware.PrometheusQueryResult_Matrix{
+							Matrix: &tripperware.Matrix{
+								SampleStreams: []tripperware.SampleStream{
+									{
+										Labels: []cortexpb.LabelAdapter{
+											{Name: "foo", Value: "bar"},
+										},
+										Samples: []cortexpb.Sample{
+											{Value: 137, TimestampMs: 1536673680000},
+											{Value: 137, TimestampMs: 1536673780000},
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+			},
 			cancelCtxBeforeDecode: true,
 			expectedDecodeErr:     context.Canceled,
+			isProtobuf:            true,
 		},
-	} {
+		{
+			promBody: &tripperware.PrometheusResponse{
+				Status: "success",
+				Data: tripperware.PrometheusData{
+					ResultType: model.ValMatrix.String(),
+					Result: tripperware.PrometheusQueryResult{
+						Result: &tripperware.PrometheusQueryResult_Matrix{
+							Matrix: &tripperware.Matrix{
+								SampleStreams: []tripperware.SampleStream{
+									{
+										Labels: []cortexpb.LabelAdapter{
+											{Name: "foo", Value: "bar"},
+										},
+										Samples: []cortexpb.Sample{
+											{Value: 137, TimestampMs: 1536673680000},
+											{Value: 137, TimestampMs: 1536673780000},
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+			cancelCtxBeforeDecode: true,
+			expectedDecodeErr:     context.Canceled,
+			isProtobuf:            false,
+		},
+	}
+	for i, tc := range testCases {
 		tc := tc
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
 			t.Parallel()
+			protobuf, err := proto.Marshal(tc.promBody)
+			require.NoError(t, err)
 			ctx, cancelCtx := context.WithCancel(context.Background())
-			response := &http.Response{
-				StatusCode: 200,
-				Header:     http.Header{"Content-Type": []string{"application/json"}},
-				Body:       io.NopCloser(bytes.NewBuffer([]byte(tc.body))),
+
+			var response *http.Response
+			if tc.isProtobuf {
+				response = &http.Response{
+					StatusCode: 200,
+					Header:     http.Header{"Content-Type": []string{tripperware.ApplicationProtobuf}},
+					Body:       io.NopCloser(bytes.NewBuffer(protobuf)),
+				}
+				tc.promBody.Headers = respHeadersProtobuf
+			} else {
+				response = &http.Response{
+					StatusCode: 200,
+					Header:     http.Header{"Content-Type": []string{tripperware.ApplicationJson}},
+					Body:       io.NopCloser(bytes.NewBuffer([]byte(tc.jsonBody))),
+				}
+				tc.promBody.Headers = respHeadersJson
 			}
+
 			if tc.cancelCtxBeforeDecode {
 				cancelCtx()
 			}
@@ -138,14 +290,15 @@ func TestResponse(t *testing.T) {
 				cancelCtx()
 				return
 			}
-			assert.Equal(t, tc.expected, resp)
+
+			assert.Equal(t, tc.promBody, resp)
 
 			// Reset response, as the above call will have consumed the body reader.
 			response = &http.Response{
 				StatusCode:    200,
 				Header:        http.Header{"Content-Type": []string{"application/json"}},
-				Body:          io.NopCloser(bytes.NewBuffer([]byte(tc.body))),
-				ContentLength: int64(len(tc.body)),
+				Body:          io.NopCloser(bytes.NewBuffer([]byte(tc.jsonBody))),
+				ContentLength: int64(len(tc.jsonBody)),
 			}
 			resp2, err := PrometheusCodec.EncodeResponse(context.Background(), resp)
 			require.NoError(t, err)
@@ -158,12 +311,13 @@ func TestResponse(t *testing.T) {
 func TestResponseWithStats(t *testing.T) {
 	t.Parallel()
 	for i, tc := range []struct {
-		body     string
-		expected *tripperware.PrometheusResponse
+		promBody   *tripperware.PrometheusResponse
+		jsonBody   string
+		isProtobuf bool
 	}{
 		{
-			body: `{"status":"success","data":{"resultType":"matrix","result":[{"metric":{"foo":"bar"},"values":[[1536673680,"137"],[1536673780,"137"]]}],"stats":{"samples":{"totalQueryableSamples":10,"totalQueryableSamplesPerStep":[[1536673680,5],[1536673780,5]],"peakSamples":16}}}}`,
-			expected: &tripperware.PrometheusResponse{
+			jsonBody: `{"status":"success","data":{"resultType":"matrix","result":[{"metric":{"foo":"bar"},"values":[[1536673680,"137"],[1536673780,"137"]]}],"stats":{"samples":{"totalQueryableSamples":10,"totalQueryableSamplesPerStep":[[1536673680,5],[1536673780,5]],"peakSamples":16}}}}`,
+			promBody: &tripperware.PrometheusResponse{
 				Status: "success",
 				Data: tripperware.PrometheusData{
 					ResultType: model.ValMatrix.String(),
@@ -196,27 +350,79 @@ func TestResponseWithStats(t *testing.T) {
 					},
 				},
 			},
+			isProtobuf: true,
+		},
+		{
+			jsonBody: `{"status":"success","data":{"resultType":"matrix","result":[{"metric":{"foo":"bar"},"values":[[1536673680,"137"],[1536673780,"137"]]}],"stats":{"samples":{"totalQueryableSamples":10,"totalQueryableSamplesPerStep":[[1536673680,5],[1536673780,5]],"peakSamples":16}}}}`,
+			promBody: &tripperware.PrometheusResponse{
+				Status: "success",
+				Data: tripperware.PrometheusData{
+					ResultType: model.ValMatrix.String(),
+					Result: tripperware.PrometheusQueryResult{
+						Result: &tripperware.PrometheusQueryResult_Matrix{
+							Matrix: &tripperware.Matrix{
+								SampleStreams: []tripperware.SampleStream{
+									{
+										Labels: []cortexpb.LabelAdapter{
+											{Name: "foo", Value: "bar"},
+										},
+										Samples: []cortexpb.Sample{
+											{Value: 137, TimestampMs: 1536673680000},
+											{Value: 137, TimestampMs: 1536673780000},
+										},
+									},
+								},
+							},
+						},
+					},
+					Stats: &tripperware.PrometheusResponseStats{
+						Samples: &tripperware.PrometheusResponseSamplesStats{
+							TotalQueryableSamples: 10,
+							TotalQueryableSamplesPerStep: []*tripperware.PrometheusResponseQueryableSamplesStatsPerStep{
+								{Value: 5, TimestampMs: 1536673680000},
+								{Value: 5, TimestampMs: 1536673780000},
+							},
+							PeakSamples: 16,
+						},
+					},
+				},
+			},
+			isProtobuf: false,
 		},
 	} {
 		tc := tc
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
 			t.Parallel()
-			tc.expected.Headers = respHeaders
-			response := &http.Response{
-				StatusCode: 200,
-				Header:     http.Header{"Content-Type": []string{"application/json"}},
-				Body:       io.NopCloser(bytes.NewBuffer([]byte(tc.body))),
+			protobuf, err := proto.Marshal(tc.promBody)
+			require.NoError(t, err)
+
+			var response *http.Response
+			if tc.isProtobuf {
+				response = &http.Response{
+					StatusCode: 200,
+					Header:     http.Header{"Content-Type": []string{tripperware.ApplicationProtobuf}},
+					Body:       io.NopCloser(bytes.NewBuffer(protobuf)),
+				}
+				tc.promBody.Headers = respHeadersProtobuf
+			} else {
+				response = &http.Response{
+					StatusCode: 200,
+					Header:     http.Header{"Content-Type": []string{tripperware.ApplicationJson}},
+					Body:       io.NopCloser(bytes.NewBuffer([]byte(tc.jsonBody))),
+				}
+				tc.promBody.Headers = respHeadersJson
 			}
+
 			resp, err := PrometheusCodec.DecodeResponse(context.Background(), response, nil)
 			require.NoError(t, err)
-			assert.Equal(t, tc.expected, resp)
+			assert.Equal(t, tc.promBody, resp)
 
 			// Reset response, as the above call will have consumed the body reader.
 			response = &http.Response{
 				StatusCode:    200,
 				Header:        http.Header{"Content-Type": []string{"application/json"}},
-				Body:          io.NopCloser(bytes.NewBuffer([]byte(tc.body))),
-				ContentLength: int64(len(tc.body)),
+				Body:          io.NopCloser(bytes.NewBuffer([]byte(tc.jsonBody))),
+				ContentLength: int64(len(tc.jsonBody)),
 			}
 			resp2, err := PrometheusCodec.EncodeResponse(context.Background(), resp)
 			require.NoError(t, err)
@@ -986,62 +1192,100 @@ func TestMergeAPIResponses(t *testing.T) {
 	}
 }
 
-func TestGzippedResponse(t *testing.T) {
+func TestCompressedResponse(t *testing.T) {
 	t.Parallel()
-	for _, tc := range []struct {
-		body   string
-		status int
-		err    error
+	for i, tc := range []struct {
+		compression string
+		jsonBody    string
+		promBody    *tripperware.PrometheusResponse
+		status      int
+		err         error
 	}{
 		{
-			body:   `{"status":"success","data":{"resultType":"matrix","result":[{"metric":{"a":"b","c":"d"},"values":[[2,"2"],[3,"3"]]}],"stats":{"samples":{"totalQueryableSamples":20,"totalQueryableSamplesPerStep":[[2,2],[3,3]],"peakSamples":10}}}}`,
-			status: 200,
+			compression: `gzip`,
+			promBody: &tripperware.PrometheusResponse{
+				Status: StatusSuccess,
+				Data: tripperware.PrometheusData{
+					ResultType: matrix,
+					Result: tripperware.PrometheusQueryResult{
+						Result: &tripperware.PrometheusQueryResult_Matrix{
+							Matrix: &tripperware.Matrix{
+								SampleStreams: []tripperware.SampleStream{
+									{
+										Labels: []cortexpb.LabelAdapter{{Name: "a", Value: "b"}, {Name: "c", Value: "d"}},
+										Samples: []cortexpb.Sample{
+											{Value: 2, TimestampMs: 2000},
+											{Value: 3, TimestampMs: 3000},
+										},
+									},
+								},
+							},
+						},
+					},
+					Stats: &tripperware.PrometheusResponseStats{Samples: &tripperware.PrometheusResponseSamplesStats{
+						TotalQueryableSamples: 20,
+						TotalQueryableSamplesPerStep: []*tripperware.PrometheusResponseQueryableSamplesStatsPerStep{
+							{Value: 2, TimestampMs: 2000},
+							{Value: 3, TimestampMs: 3000},
+						},
+						PeakSamples: 10,
+					}},
+				},
+				Headers: []*tripperware.PrometheusResponseHeader{},
+			},
+			jsonBody: `{"status":"success","data":{"resultType":"matrix","result":[{"metric":{"a":"b","c":"d"},"values":[[2,"2"],[3,"3"]]}],"stats":{"samples":{"totalQueryableSamples":20,"totalQueryableSamplesPerStep":[[2,2],[3,3]],"peakSamples":10}}}}`,
+			status:   200,
 		},
 		{
-			body:   `error generic 400`,
-			status: 400,
-			err:    httpgrpc.Errorf(400, `error generic 400`),
+			compression: `gzip`,
+			jsonBody:    `error generic 400`,
+			status:      400,
+			err:         httpgrpc.Errorf(400, `error generic 400`),
 		},
 		{
-			status: 400,
-			err:    httpgrpc.Errorf(400, ""),
+			compression: `gzip`,
+			status:      400,
+			err:         httpgrpc.Errorf(400, ""),
 		},
 	} {
-		for _, c := range []bool{true, false} {
-			c := c
-			t.Run(fmt.Sprintf("compressed %t [%s]", c, tc.body), func(t *testing.T) {
-				t.Parallel()
-				h := http.Header{
-					"Content-Type": []string{"application/json"},
-				}
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			t.Parallel()
+			h := http.Header{}
+			var b []byte
+			if tc.promBody != nil {
+				protobuf, err := proto.Marshal(tc.promBody)
+				b = protobuf
+				require.NoError(t, err)
+				h.Set("Content-Type", tripperware.ApplicationProtobuf)
+				tc.promBody.Headers = append(tc.promBody.Headers, &tripperware.PrometheusResponseHeader{Name: "Content-Type", Values: []string{tripperware.ApplicationProtobuf}})
+			} else {
+				b = []byte(tc.jsonBody)
+				h.Set("Content-Type", tripperware.ApplicationJson)
+			}
 
-				responseBody := bytes.NewBuffer([]byte(tc.body))
-				if c {
-					h.Set("Content-Encoding", "gzip")
-					var buf bytes.Buffer
-					w := gzip.NewWriter(&buf)
-					_, err := w.Write([]byte(tc.body))
-					require.NoError(t, err)
-					w.Close()
-					responseBody = &buf
-				}
+			h.Set("Content-Encoding", tc.compression)
+			if tc.promBody != nil {
+				tc.promBody.Headers = append(tc.promBody.Headers, &tripperware.PrometheusResponseHeader{Name: "Content-Encoding", Values: []string{"gzip"}})
+			}
+			responseBody := &bytes.Buffer{}
+			w := gzip.NewWriter(responseBody)
+			_, err := w.Write(b)
+			require.NoError(t, err)
+			w.Close()
 
-				response := &http.Response{
-					StatusCode: tc.status,
-					Header:     h,
-					Body:       io.NopCloser(responseBody),
-				}
-				r, err := PrometheusCodec.DecodeResponse(context.Background(), response, nil)
-				require.Equal(t, tc.err, err)
+			response := &http.Response{
+				StatusCode: tc.status,
+				Header:     h,
+				Body:       io.NopCloser(responseBody),
+			}
+			resp, err := PrometheusCodec.DecodeResponse(context.Background(), response, nil)
+			require.Equal(t, tc.err, err)
 
-				if err == nil {
-					resp, err := json.Marshal(r)
-					require.NoError(t, err)
-
-					require.Equal(t, tc.body, string(resp))
-				}
-			})
-		}
+			if err == nil {
+				require.NoError(t, err)
+				require.Equal(t, tc.promBody.Data, resp.(*tripperware.PrometheusResponse).Data)
+			}
+		})
 	}
 }